chore(server): move types
