+++ packages/twenty-server/test/integration/graphql/types/common-operation-input.type.ts
