+++ packages/twenty-server/test/integration/graphql/types/common-response-body.type.ts
