+++ packages/twenty-server/test/integration/graphql/utils/activate-workspace.util.ts
@@ -1,7 +1,7 @@
 import { ActivateWorkspaceInput } from 'src/engine/core-modules/workspace/dtos/activate-workspace-input';
+import { CommonOperationInput } from 'test/integration/graphql/types/common-operation-input.type';
 import { activateWorkspaceOperationFactory } from 'test/integration/graphql/utils/activate-workspace-factory.util';
 import { makeMetadataAPIRequest } from 'test/integration/graphql/utils/make-metadata-api-request.util';
-import { CommonOperationInput } from 'test/integration/types/common-operation-input.type';
 
 export const performActivateWorkspace = async ({
   input,
