+++ packages/twenty-server/test/integration/metadata/suites/field-metadata/utils/delete-one-field-metadata-query-factory.util.ts
@@ -1,5 +1,5 @@
 import gql from 'graphql-tag';
-import { PerformMetadataQueryParams } from 'test/integration/types/perform-metadata-query.type';
+import { PerformMetadataQueryParams } from 'test/integration/graphql/types/perform-metadata-query.type';
 
 export type DeleteOneFieldFactoryInput = {
   idToDelete: string;
