+++ packages/twenty-server/test/integration/metadata/suites/object-metadata/utils/create-one-object-metadata-query-factory.util.ts
@@ -1,5 +1,5 @@
 import gql from 'graphql-tag';
-import { PerformMetadataQueryParams } from 'test/integration/types/perform-metadata-query.type';
+import { PerformMetadataQueryParams } from 'test/integration/graphql/types/perform-metadata-query.type';
 
 import { CreateObjectInput } from 'src/engine/metadata-modules/object-metadata/dtos/create-object.input';
